Edit README.md to reflect the real function of the program
I had set out to make this program a full Python-based chat program.
Real life got in the way before it got further than an RSA demo.  I will
not continue work on it right now, but I wanted to correct the readme so
it says what the program really is right now.